add upgradable contract and deploy script
