--- deploy/pegged/tokens/008_upgradable_multi_bridge_token.ts
@@ -1,28 +1,0 @@
-import * as dotenv from 'dotenv';
-import { DeployFunction } from 'hardhat-deploy/types';
-import { HardhatRuntimeEnvironment } from 'hardhat/types';
-
-dotenv.config();
-
-const deployFunc: DeployFunction = async (hre: HardhatRuntimeEnvironment) => {
-  const { deployments, getNamedAccounts } = hre;
-  const { deploy } = deployments;
-  const { deployer } = await getNamedAccounts();
-
-  await deploy('MultiBridgeToken', {
-    from: deployer,
-    log: true,
-    args: [
-      process.env.MULTI_BRIDGE_TOKEN_NAME,
-      process.env.MULTI_BRIDGE_TOKEN_SYMBOL,
-      process.env.MULTI_BRIDGE_TOKEN_DECIMALS
-    ],
-    proxy: {
-      proxyContract: "OptimizedTransparentProxy"
-    }
-  });
-};
-
-deployFunc.tags = ['UpgradableMultiBridgeToken'];
-deployFunc.dependencies = [];
-export default deployFunc;
